requirements.txt
requests
beautifulsoup4
openpyxl
pillow
flask
pandas
